# Optimize nap times while the runner waiting for a vm